refactor(frontend): AppChrome を widgets→features/app-chrome へ移設する（#284・R1① 5層固定）
hub 差し戻し: src/widgets/ 新設は R1①(会議① FSD 5層固定・widgets/processes
新設禁止・src 直下は app/pages/features/entities/shared のみ)違反。eslint 合成形
の zones は層存在を検査しないため黙るが depcruise/C1 で赤くなる偽緑。

AppChrome は『i18n を束ねて shared/ui AppShell に prop を配る』= feature の仕事。
features/app-chrome/ui/AppChrome へ移し、公開バレル features/app-chrome/index.ts
経由で 7 pages が import(pages→features 正方向・features は i18n 使用可)。

再実測: 合成形 no-restricted-imports=0 維持・import-x は fork B の2件のみ(新違反
なし)・type-check/lint/test 272/format/ratchet 緑。
