+++ frontend/src/features/app-chrome/index.ts
@@ -1,0 +1,1 @@
+export { AppChrome } from './ui/AppChrome';
